[WIP]debug for humble
