+++ helper_scripts/generate_messages_model_helper.sh
