+++ helper_scripts/messages_generator_runner.sh
